S104 split: terraform_vars + ngfw_polling
